Fix crash when daemon is down

+++ src/remote-wipe-service.cpp
@@ -100,6 +100,10 @@ void RemoteWipeService::onRequestSuccess()
 
 void RemoteWipeService::askDaemonDeleteAccount(const Account& account)
 {
+    if (!gui->rpcClient()->isConnected()) {
+        qWarning() << "Failed to remove local cache of account" << account;
+        return;
+    }
     if (!gui->rpcClient()->deleteAccount(account, true)) {
         qWarning() << "Failed to remove local cache of account" << account;
     }
